+++ lib/db/gql-pubsub-sequelize-engine.d.ts
@@ -1,9 +1,8 @@
-import { PubSubEngine } from 'type-graphql';
-import { Sequelize } from 'sequelize-typescript';
-import { Model } from 'sequelize';
+import { RfiPubSubEngine } from '../pubsub';
+import { Model, Sequelize } from 'sequelize-typescript';
 /**
  *
  */
-export declare function linkSequelizeToPubSubEngine(pubSub: PubSubEngine, sequelize: Sequelize): void;
+export declare function linkSequelizeToPubSubEngine(pubSub: RfiPubSubEngine, sequelize: Sequelize): void;
 export declare function publishCurrentState(instance: Model<any, any>): void;
-export declare function pubSubFrom(sequelize: Sequelize): PubSubEngine | null;
+export declare function pubSubFrom(sequelize: Sequelize): RfiPubSubEngine | null;