+++ lib/gql/node.resolver.d.ts
@@ -1,7 +1,8 @@
 import { PubSubEngine } from 'type-graphql';
 import { Node, RelayResolver } from './index';
 import { NodeService } from './relay.service';
-import { NodeNotification, NotificationOf, DbModelChangeNotification } from './node-notification';
+import { NodeNotification, NotificationOf } from './node-notification';
+import { RawPayload } from '../pubsub/helper';
 declare class ClassGqlNodeNotification extends NodeNotification<any> {
     sequence: number;
     notificationOf: NotificationOf;
@@ -13,7 +14,7 @@ export declare class NodeResolver implements RelayResolver {
     constructor(nodeServices: Array<NodeService<any>>);
     node(oidString: string, ctx: any): Promise<Node<any> | null>;
     publishLastKnownState(oidString: string, pubSub: PubSubEngine, ctx: any): boolean;
-    onChange(payload: DbModelChangeNotification): Promise<ClassGqlNodeNotification>;
+    onChange(rawPayload: RawPayload): Promise<ClassGqlNodeNotification>;
     unWrapOid(oidString: string, ctx: any): Promise<string>;
     makeOid(scope: string, id: string, ctx: any): Promise<string>;
 }