+++ src/db/gql-pubsub-sequelize-engine.ts
@@ -1,17 +1,12 @@
-import { PubSubEngine } from 'type-graphql';
-import { Sequelize } from 'sequelize-typescript';
-import { Model, CreateOptions, UpdateOptions } from 'sequelize';
-import {
-  NotificationOf,
-  DbModelChangeNotification,
-  NODE_CHANGE_NOTIFICATION,
-  ModelDelta
-} from '../gql/node-notification';
+import { RfiPubSubEngine } from '../pubsub';
+import { Model, Sequelize } from 'sequelize-typescript';
+import { CreateOptions, UpdateOptions } from 'sequelize';
+import { NotificationOf, ModelDelta } from '../gql/node-notification';
 
 /**
  *
  */
-export function linkSequelizeToPubSubEngine(pubSub: PubSubEngine, sequelize: Sequelize) {
+export function linkSequelizeToPubSubEngine(pubSub: RfiPubSubEngine, sequelize: Sequelize) {
   // Install hooks on Sequelize that publish GqlNodeNotifications
   // Takes advantage of the RFI frameworks connection of sequelize Model<> class to our
   // GQL classes
@@ -20,22 +15,25 @@ export function linkSequelizeToPubSubEngine(pubSub: PubSubEngine, sequelize: Seq
   sequelize.afterCreate((instance, options) => {
     // Cache the previous values to closure so they can be published __after__ commit
     // (the previous vales will be overwritten on commit)
-    const deltas = getChangedAttributes(instance);
+    // NOTE there is a slight incompatibility between the sequelize-typescript lib and sequelize in the type
+    // definition of the instance passed to the hooks... It actually doesnt matter for now, so we assert to any
+    //
+    const deltas = getChangedAttributes(instance as any);
 
     if (options && options.transaction) {
-      options.transaction.afterCommit(t => gqlCreateHook(pubSub, instance, deltas, options));
+      options.transaction.afterCommit(t => gqlCreateHook(pubSub, instance as any, deltas, options));
       return;
     }
-    gqlCreateHook(pubSub, instance, deltas, options);
+    gqlCreateHook(pubSub, instance as any, deltas, options);
   });
 
   sequelize.afterUpdate((instance, options) => {
-    const deltas = getChangedAttributes(instance);
+    const deltas = getChangedAttributes(instance as any);
     if (options && options.transaction) {
-      options.transaction.afterCommit(t => gqlUpdateHook(pubSub, instance, deltas, options));
+      options.transaction.afterCommit(t => gqlUpdateHook(pubSub, instance as any, deltas, options));
       return;
     }
-    gqlUpdateHook(pubSub, instance, deltas, options);
+    gqlUpdateHook(pubSub, instance as any, deltas, options);
   });
   // sequelize.afterBulkCreate((instances, options) => gqlBulkCreateHook(pubSub, instances, options));
 }
@@ -61,22 +59,17 @@ function getChangedAttributes(instance: Model<any, any>): ModelDelta[] {
 export function publishCurrentState(instance: Model<any, any>) {
   const pubSub = pubSubFrom(instance.sequelize as Sequelize);
   if (pubSub) {
-    const payload = new DbModelChangeNotification(NotificationOf.LAST_KNOWN_STATE, instance, []);
-    // This should be its own promise chain
-    // tslint:disable-next-line: no-floating-promises
-    pubSub.publish(NODE_CHANGE_NOTIFICATION, payload);
-    // Also publish the specific Model
-    // This should be its own promise chain
-    // tslint:disable-next-line: no-floating-promises
-    pubSub.publish(`${NODE_CHANGE_NOTIFICATION}_${instance.constructor.name}`, payload);
+    pubSub.publishPayload(NotificationOf.LAST_KNOWN_STATE, instance, []);
   }
 }
 
+// It would not compile/run if I moved these under pubsub
 const PubSubKey = Symbol('PubSubEngine');
-function attachPubSubEngineToSequelize(pubSub: PubSubEngine, sequelize: Sequelize): void {
+function attachPubSubEngineToSequelize(pubSub: RfiPubSubEngine, sequelize: Sequelize): void {
   Reflect.set(sequelize, PubSubKey, pubSub);
 }
-export function pubSubFrom(sequelize: Sequelize): PubSubEngine | null {
+
+export function pubSubFrom(sequelize: Sequelize): RfiPubSubEngine | null {
   const pubSub = Reflect.get(sequelize, PubSubKey);
   return pubSub ? pubSub : null;
 }
@@ -89,32 +82,18 @@ export function pubSubFrom(sequelize: Sequelize): PubSubEngine | null {
 
 // The Resolvers will convert to a Gql subscription
 function gqlCreateHook(
-  pubSub: PubSubEngine,
+  pubSub: RfiPubSubEngine,
   instance: Model<any, any>,
   deltas: ModelDelta[],
   options: CreateOptions
 ) {
-  const payload = new DbModelChangeNotification(NotificationOf.CREATED, instance, deltas);
-  // This should be its own promise chain
-  // tslint:disable-next-line: no-floating-promises
-  pubSub.publish(NODE_CHANGE_NOTIFICATION, payload);
-  // Also publish the specific Model
-  // This should be its own promise chain
-  // tslint:disable-next-line: no-floating-promises
-  pubSub.publish(`${NODE_CHANGE_NOTIFICATION}_${instance.constructor.name}`, payload);
+  pubSub.publishPayload(NotificationOf.CREATED, instance, deltas);
 }
 function gqlUpdateHook(
-  pubSub: PubSubEngine,
+  pubSub: RfiPubSubEngine,
   instance: Model<any, any>,
   deltas: ModelDelta[],
   options: UpdateOptions
 ) {
-  const payload = new DbModelChangeNotification(NotificationOf.UPDATED, instance, deltas);
-  // This should be its own promise chain
-  // tslint:disable-next-line: no-floating-promises
-  pubSub.publish(NODE_CHANGE_NOTIFICATION, payload);
-  // Also publish the specific Model
-  // This should be its own promise chain
-  // tslint:disable-next-line: no-floating-promises
-  pubSub.publish(`${NODE_CHANGE_NOTIFICATION}_${instance.constructor.name}`, payload);
+  pubSub.publishPayload(NotificationOf.UPDATED, instance, deltas);
 }